+++ platform/nxp/lpc13xx/flash.c
@@ -85,7 +85,7 @@ static enum Result flashGetParam(void *object, enum IfParameter parameter,
   switch ((enum FlashParameter)parameter)
   {
     case IF_FLASH_PAGE_SIZE:
-      *(size_t *)data = FLASH_PAGE_SIZE;
+      *(uint32_t *)data = FLASH_PAGE_SIZE;
       return E_OK;
 
     default:
@@ -95,11 +95,11 @@ static enum Result flashGetParam(void *object, enum IfParameter parameter,
   switch (parameter)
   {
     case IF_POSITION:
-      *(size_t *)data = interface->position;
+      *(uint32_t *)data = interface->position;
       return E_OK;
 
     case IF_SIZE:
-      *(size_t *)data = interface->size;
+      *(uint32_t *)data = interface->size;
       return E_OK;
 
     default:
@@ -117,7 +117,7 @@ static enum Result flashSetParam(void *object, enum IfParameter parameter,
   {
     case IF_FLASH_ERASE_SECTOR:
     {
-      const size_t position = *(const size_t *)data;
+      const uintptr_t position = *(const uint32_t *)data;
 
       if (!isSectorPositionValid(interface, position))
         return E_ADDRESS;
@@ -136,7 +136,7 @@ static enum Result flashSetParam(void *object, enum IfParameter parameter,
   {
     case IF_POSITION:
     {
-      const size_t position = *(const size_t *)data;
+      const uintptr_t position = *(const uint32_t *)data;
 
       if (position < interface->size)
       {
